pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gastown"
version = "0.1.0"
description = "Multi-agent collaborative development system using git"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
gastown = "gastown.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["gastown"]
